refactor(store): align submitted prompts beside message records
