+++ crates/rimz/src/store/message.rs
@@ -63,6 +63,16 @@ pub enum HarnessNotice {
     Other(String),
 }
 
+impl HarnessNotice {
+    pub(crate) fn header_type(&self) -> String {
+        match self {
+            Self::SubagentReport => "SUBAGENT_REPORT".to_owned(),
+            Self::Wake => "WAKE".to_owned(),
+            Self::Other(notice) => notice.to_ascii_uppercase(),
+        }
+    }
+}
+
 impl MessageSender {
     pub fn render(&self) -> String {
         match self {
@@ -859,6 +869,115 @@ pub fn claim_expired(last_attempt_at: Option<Timestamp>, now: Timestamp) -> bool
     };
     let age = now.duration_since(last);
     age.is_negative() || (age.as_secs() as u64) >= CLAIM_TTL.as_secs()
+}
+
+/// Align one submitted pane paste with the records written as its batch.
+///
+/// Record text supplies the otherwise ambiguous boundaries between adjacent
+/// messages. Agent-, subagent-, and human-authored records consume their
+/// rendered headers; system records stay verbatim. Interior record whitespace
+/// matches verbatim; only the paste's outer first/last whitespace follows hook
+/// normalization. Composer text around an intact batch is returned separately.
+pub fn align_submitted_prompt<'a>(
+    prompt: &'a str,
+    records: &[&MessageRecord],
+) -> Option<(Option<&'a str>, Vec<&'a str>, Option<&'a str>)> {
+    if records.is_empty() {
+        return None;
+    }
+    let prompt = prompt.trim();
+    let anchor = match &records[0].sender {
+        MessageSender::Agent { .. } => "Type: AGENT_MESSAGE\nFrom: @".to_owned(),
+        MessageSender::Subagent { .. } => "Type: SUBAGENT_REPORT\nFrom: @".to_owned(),
+        MessageSender::Harness { notice } => {
+            format!("Type: {}\nFrom: @rimz\nContent:\n", notice.header_type())
+        }
+        MessageSender::Human => "Type: USER_MESSAGE\nFrom: @user\nContent:\n".to_owned(),
+        MessageSender::System => {
+            let (segments, trailing) = align_submitted_prompt_from(prompt, records)?;
+            return trailing.is_empty().then_some((None, segments, None));
+        }
+    };
+    for (start, _) in prompt.match_indices(&anchor) {
+        let Some((segments, trailing)) = align_submitted_prompt_from(&prompt[start..], records)
+        else {
+            continue;
+        };
+        return Some((
+            nonempty_trimmed(&prompt[..start]),
+            segments,
+            nonempty_trimmed(trailing),
+        ));
+    }
+    None
+}
+
+fn align_submitted_prompt_from<'a>(
+    prompt: &'a str,
+    records: &[&MessageRecord],
+) -> Option<(Vec<&'a str>, &'a str)> {
+    let mut remaining = prompt;
+    let mut segments = Vec::with_capacity(records.len());
+    for (index, record) in records.iter().enumerate() {
+        let segment = remaining;
+        let mut body = segment;
+        match &record.sender {
+            MessageSender::Agent { .. } => {
+                let attributed = body.strip_prefix("Type: AGENT_MESSAGE\nFrom: @")?;
+                let (handle, text) = attributed.split_once("\nContent:\n")?;
+                if handle.is_empty() || handle.chars().any(char::is_whitespace) {
+                    return None;
+                }
+                body = text;
+            }
+            MessageSender::Subagent { .. } => {
+                let attributed = body.strip_prefix("Type: SUBAGENT_REPORT\nFrom: @")?;
+                let (handle, text) = attributed.split_once("\nContent:\n")?;
+                if handle.is_empty() || handle.chars().any(char::is_whitespace) {
+                    return None;
+                }
+                body = text;
+            }
+            MessageSender::Harness { notice } => {
+                let header = format!("Type: {}\nFrom: @rimz\nContent:\n", notice.header_type());
+                body = body.strip_prefix(&header)?;
+            }
+            MessageSender::Human => {
+                body = body.strip_prefix("Type: USER_MESSAGE\nFrom: @user\nContent:\n")?;
+            }
+            MessageSender::System => {}
+        }
+        let first = index == 0;
+        let last = index + 1 == records.len();
+        if first {
+            body = body.trim_start();
+        }
+        let expected = match (first, last) {
+            (true, true) => record.text.trim(),
+            (true, false) => record.text.trim_start(),
+            (false, true) => record.text.trim_end(),
+            (false, false) => record.text.as_str(),
+        };
+        if expected.is_empty() {
+            return None;
+        }
+        let after_text = body.strip_prefix(expected)?;
+        if last {
+            let segment_end = segment.len() - after_text.len();
+            segments.push(segment[..segment_end].trim());
+            return Some((segments, after_text));
+        }
+        let after_join = after_text.strip_prefix("\n\n")?;
+        let segment_end = segment.len() - after_text.len();
+        segments.push(segment[..segment_end].trim());
+        remaining = after_join;
+    }
+    None
+}
+
+fn nonempty_trimmed(text: &str) -> Option<&str> {
+    let text = text.trim();
+    (!text.is_empty()).then_some(text)
 }
 
 #[cfg(test)]
