+++ crates/rimz/src/store/message/tests.rs
@@ -838,3 +838,121 @@ pub(crate) fn when_condition(
         met_at,
     }
 }
+
+#[test]
+fn align_submitted_prompt_consumes_human_header() {
+    let recipient = agent("session-recipient", None);
+    let record = MessageRecord::new(
+        WorkspaceId::from_project_root(std::path::Path::new("/tmp/rimz-target-test")),
+        &recipient,
+        "ship it".to_owned(),
+        true,
+        crate::store::message::DeliveryGate::Done,
+    );
+    let prompt = "Type: USER_MESSAGE\nFrom: @user\nContent:\nship it";
+
+    let (leading, segments, trailing) =
+        align_submitted_prompt(prompt, &[&record]).expect("aligned prompt");
+    assert_eq!((leading, trailing), (None, None));
+    assert_eq!(segments, vec![prompt]);
+}
+
+#[test]
+fn align_submitted_prompt_consumes_harness_report_header() {
+    let recipient = agent("session-recipient", None);
+    let record = MessageRecord::new(
+        WorkspaceId::from_project_root(std::path::Path::new("/tmp/rimz-target-test")),
+        &recipient,
+        "ship it".to_owned(),
+        true,
+        crate::store::message::DeliveryGate::Done,
+    )
+    .with_sender(MessageSender::Harness {
+        notice: HarnessNotice::SubagentReport,
+    });
+    let prompt = "Type: SUBAGENT_REPORT\nFrom: @rimz\nContent:\nship it";
+
+    let (leading, segments, trailing) =
+        align_submitted_prompt(prompt, &[&record]).expect("aligned prompt");
+    assert_eq!((leading, trailing), (None, None));
+    assert_eq!(segments, vec![prompt]);
+}
+
+#[test]
+fn align_submitted_prompt_separates_stray_composer_text() {
+    let recipient = agent("session-recipient", None);
+    let first = MessageRecord::new(
+        WorkspaceId::from_project_root(std::path::Path::new("/tmp/rimz-target-test")),
+        &recipient,
+        "first".to_owned(),
+        true,
+        crate::store::message::DeliveryGate::Done,
+    );
+    let second = MessageRecord::new(
+        first.workspace_id.clone(),
+        &recipient,
+        "second".to_owned(),
+        true,
+        crate::store::message::DeliveryGate::Done,
+    );
+    let first_prompt = "Type: USER_MESSAGE\nFrom: @user\nContent:\nfirst";
+    let second_prompt = "Type: USER_MESSAGE\nFrom: @user\nContent:\nsecond";
+    let prompt = format!(
+        "decoy Type: USER_MESSAGE\nFrom: @user\nContent:\nnot it before{first_prompt}\n\n{second_prompt}after"
+    );
+
+    let (leading, segments, trailing) =
+        align_submitted_prompt(&prompt, &[&first, &second]).expect("embedded batch aligned");
+    assert_eq!(
+        leading,
+        Some("decoy Type: USER_MESSAGE\nFrom: @user\nContent:\nnot it before")
+    );
+    assert_eq!(segments, vec![first_prompt, second_prompt]);
+    assert_eq!(trailing, Some("after"));
+    assert_eq!(leading.map_or(0, str::len), 59);
+    assert_eq!(trailing.map_or(0, str::len), 5);
+}
+
+#[test]
+fn align_submitted_prompt_rejects_a_truncated_record() {
+    let recipient = agent("session-recipient", None);
+    let record = MessageRecord::new(
+        WorkspaceId::from_project_root(std::path::Path::new("/tmp/rimz-target-test")),
+        &recipient,
+        "ship it".to_owned(),
+        true,
+        crate::store::message::DeliveryGate::Done,
+    );
+
+    assert_eq!(
+        align_submitted_prompt(
+            "Type: USER_MESSAGE\nFrom: @user\nContent:\nship",
+            &[&record]
+        ),
+        None
+    );
+}
+
+#[test]
+fn align_submitted_prompt_requires_exact_system_text() {
+    let recipient = agent("session-recipient", None);
+    let record = MessageRecord::new(
+        WorkspaceId::from_project_root(std::path::Path::new("/tmp/rimz-target-test")),
+        &recipient,
+        "continue".to_owned(),
+        true,
+        crate::store::message::DeliveryGate::Done,
+    )
+    .with_sender(MessageSender::System);
+
+    assert_eq!(
+        align_submitted_prompt("continue", &[&record]),
+        Some((None, vec!["continue"], None))
+    );
+    for prompt in [
+        "continue with the refactor",
+        "please continue where you left off",
+    ] {
+        assert_eq!(align_submitted_prompt(prompt, &[&record]), None, "{prompt}");
+    }
+}
