+++ docs/internals/harness/messaging.md
@@ -30,7 +30,7 @@ Start here when you are looking for where a behaviour lives.
 | [`message/deliver.rs`](../../../crates/rimz/src/message/deliver.rs) | Readiness: the ordered delivery check, the delivery attempt and its failure recovery, condition evaluation, the sweep, and the wake stamp. |
 | [`message/reply.rs`](../../../crates/rimz/src/message/reply.rs) | `--wait`: leg state machines, transcript anchoring, cycle detection, join settlement. |
 | [`message/fire.rs`](../../../crates/rimz/src/message/fire.rs) | The elder's side of the clock: read the wake stamp, spawn `message sweep`. Nothing else. |
-| [`store/message.rs`](../../../crates/rimz/src/store/message.rs) | The durable record schema and vocabulary, card matching, and FIFO/claim/batch selection. |
+| [`store/message.rs`](../../../crates/rimz/src/store/message.rs) | The durable record schema and vocabulary, card matching, FIFO/claim/batch selection, and the prompt alignment [confirmation](#confirmation-and-retry) runs against those records. |
 | [`store/message/codec.rs`](../../../crates/rimz/src/store/message/codec.rs) | The JSONL live queue and terminal history codec. |
 | [`store/writer/queue.rs`](../../../crates/rimz/src/store/writer/queue.rs) | Every status transition, under the workspace lock, with its audit event. |
 | [`cli/message/`](../../../crates/rimz/src/cli/message) | Flag parsing, rendering, and the inbox verbs. No delivery logic. |
@@ -241,7 +241,7 @@ The batch lands as one paste and one submit. Agent- and human-authored members k
 
 ### Confirmation and retry
 
-A `TurnStarted` hook aligns the submitted prompt against candidate `Prompt` batches for that card. Durable record text supplies the boundaries between adjacent messages; agent- and human-authored records consume their structured headers, while system records match verbatim. Record text and the blank-line join match verbatim inside a batch; only the first record's leading and last record's trailing whitespace follow the hook payload's outer normalization. A submission that contains a whole headered batch confirms it even when composer text appears before or after it. That stray text becomes separate direct-input transcript entries, credits the human's input, and the delivered event records a mixed-submit reason with the stray byte counts. Stray composer text is never a reason to write the pane again. Headerless system batches require an exact whole-prompt match because their text alone cannot distinguish a delivery from the user's own words. Reported text that contains no intact candidate batch is direct pane input and confirms nothing.
+A `TurnStarted` hook aligns the submitted prompt against candidate `Prompt` batches for that card. [`align_submitted_prompt`](../../../crates/rimz/src/store/message.rs) does the matching, beside the record vocabulary it reads. Durable record text supplies the boundaries between adjacent messages; agent- and human-authored records consume their structured headers, while system records match verbatim. Record text and the blank-line join match verbatim inside a batch; only the first record's leading and last record's trailing whitespace follow the hook payload's outer normalization. A submission that contains a whole headered batch confirms it even when composer text appears before or after it. That stray text becomes separate direct-input transcript entries, credits the human's input, and the delivered event records a mixed-submit reason with the stray byte counts. Stray composer text is never a reason to write the pane again. Headerless system batches require an exact whole-prompt match because their text alone cannot distinguish a delivery from the user's own words. Reported text that contains no intact candidate batch is direct pane input and confirms nothing.
 
 When a turn-start adapter reports no usable prompt text, confirmation falls back to the oldest `Sent` prompt and its `batch_id`, preserving hookless-text compatibility. A `Compacting` hook uses the same oldest-`Sent` fallback for `Command` records. Correlation never selects a `Claimed` record because its deliverer owns an in-progress pane write.
