+++ refactor-target.toml
@@ -244,7 +244,7 @@ upward-dependencies = [
     "harness::target",
     "remote::link",
 ]
-surface-budget = 314
+surface-budget = 316
 
 [[module]]
 path = "crates/rimz/src/theme"
